+++ src/components/AppRoutes.js
@@ -1,5 +1,7 @@
 import { React } from 'react'
 import { Route, Routes, Navigate, useParams } from 'react-router-dom'
+import * as Sentry from "@sentry/react";
+
 import ProjectComponentLoader from './Editor/ProjectComponentLoader/ProjectComponentLoader'
 import ProjectIndex from './ProjectIndex/ProjectIndex'
 import EmbeddedViewer from './EmbeddedViewer/EmbeddedViewer'
@@ -14,8 +16,10 @@ const ProjectsRedirect = () => {
   return <Navigate replace to={`/en/projects/${identifier}`} />
 }
 
+const SentryRoutes = Sentry.withSentryReactRouterV6Routing(Routes)
+
 const AppRoutes = () => (
-  <Routes>
+  <SentryRoutes>
     <Route
       path="/auth/callback"
       element={<Callback/>}
@@ -48,7 +52,7 @@ const AppRoutes = () => (
     { localeRedirects.map(link => {
       return <Route key={link} path={link} element={<Navigate replace to={`/en${link}`} />} />
     }) }
-  </Routes>
+  </SentryRoutes>
 )
 
 export default AppRoutes