+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Updated privacy policy link to use the child friendly privacy policy (#397)
 - Update URL structure to include locale (#407)
+- Update Sentry configuration to allow distributed tracing (#411)
 
 ### Fixed